pyproject.toml
[tool.poetry]
name = "interview_ct"
version = "0.1.0"
description = ""
authors = ["Weslly Oliveira <weslly99@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
python-dotenv = "^0.20.0"
Flask-RESTful = "^0.3.9"
requests = "^2.27.1"
redis = "^4.3.3"
Flask-Caching = "^1.11.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
